cloud_storage: Move materialized_manifest_cache
Move materialized_manifest_cache into a dedicated .cc to avoid cyclic
dependency.
