+++ src/v/cloud_storage/CMakeLists.txt
@@ -37,6 +37,7 @@ v_cc_library(
     segment_chunk_api.cc
     segment_chunk_data_source.cc
     async_manifest_view.cc
+    materialized_manifest_cache.cc
   DEPS
     Seastar::seastar
     v::bytes
