+++ src/v/cloud_storage/fwd.h
@@ -21,6 +21,7 @@ class partition_manifest;
 class topic_manifest;
 class partition_probe;
 class async_manifest_view;
+class partition_probe;
 
 struct log_recovery_result;
 struct offset_range;
